requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
Brotli @ file:///D:/bld/brotli-split_1648883782420/work
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
colorama==0.4.6
cycler==0.11.0
Cython==3.0.6
edge-tts==6.1.10
ffmpeg-python==0.2.0
ffpyplayer==4.5.1
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
gevent==22.10.2
google-auth==2.25.2
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.60.0
huggingface-hub==0.16.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imgviz==1.7.4
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==9.1.0
multidict==6.0.5
numba==0.56.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openai-whisper==20230124
packaging==23.1
pandas==1.3.5
Pillow @ file:///D:/bld/pillow_1621288212302/work
playsound2==1.3.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio @ file:///C:/Users/Administrator/Downloads/%E8%AE%BA%E6%96%87/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycocotools-windows==2.0.0.2
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks @ file:///D:/bld/pysocks_1648857448477/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets @ file:///C:/ci/pywavelets_1648728036674/work
PyYAML==6.0.1
regex==2023.12.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.0.2
scipy @ file:///C:/ci/scipy_1616703425152/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stable-ts==1.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.1
transformers==4.30.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1695228328709/work
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==2.2.3
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wincertstore==0.2
yarl==1.9.4
zipp==3.15.0
zope.event==5.0
zope.interface==6.2
